Module Client.Rpc

type 'a handler = [ `write ] H2.Body.t Lwt.t -> [ `read ] H2.Body.t -> 'a Lwt.t

val bidirectional_streaming : f:((Pbrt.Encoder.t -> unit) -> Pbrt.Decoder.t Lwt_stream.t -> 'a Lwt.t) -> 'a handler

bidirectional_streaming ~f write read sets up the sending and receiving logic using write and read, then calls f with a push function for requests and a stream of responses.

val client_streaming : f:((Pbrt.Encoder.t -> unit) -> Pbrt.Decoder.t option Lwt.t -> 'a Lwt.t) -> 'a handler

client_streaming ~f write read sets up the sending and receiving logic using write and read, then calls f with a push function for requests and promise for the response.

val server_streaming : f:(Pbrt.Decoder.t Lwt_stream.t -> 'a Lwt.t) -> Pbrt.Encoder.t -> 'a handler

server_streaming ~f enc write read sets up the sending and receiving logic using write and read, then sends enc and calls f with a stream of responses.

val unary : f:(Pbrt.Decoder.t option Lwt.t -> 'a Lwt.t) -> Pbrt.Encoder.t -> 'a handler

unary ~f enc write read sets up the sending and receiving logic using write and read, then sends enc and calls f with a promise for the response.